refactor(router-core): begin migrating types from react-router and solid-router to router-core
