+++ packages/react-router/src/router.ts
@@ -46,6 +46,7 @@ import type {
   CommitLocationOptions,
   ControlledPromise,
   Manifest,
+  NavigateOptions,
   NonNullableUpdater,
   ParsedLocation,
   PickAsRequired,
@@ -55,16 +56,18 @@ import type {
   SearchSerializer,
   StartSerializer,
   TrailingSlashOption,
+  ToOptions,
   Updater,
   ViewTransitionOptions,
+  Register,
+  MakeRemountDepsOptionsUnion,
+  AnyRouteWithContext,
 } from '@tanstack/router-core'
 import type {
   AnyRoute,
-  AnyRouteWithContext,
   BeforeLoadContextOptions,
   ErrorRouteComponent,
   LoaderFnContext,
-  MakeRemountDepsOptionsUnion,
   NotFoundRouteComponent,
   RootRoute,
   RouteComponent,
@@ -78,7 +81,7 @@ import type {
   RoutePaths,
   RoutesById,
   RoutesByPath,
-} from './routeInfo'
+} from '@tanstack/router-core'
 import type {
   AnyRouteMatch,
   MakeRouteMatch,
@@ -86,11 +89,10 @@ import type {
   MatchRouteOptions,
 } from './Matches'
 
-import type { BuildLocationFn, NavigateFn } from './RouterProvider'
+import type { BuildLocationFn, NavigateFn } from '@tanstack/router-core'
 
 import type { AnyRedirect, ResolvedRedirect } from './redirects'
 import type { NotFoundError } from './not-found'
-import type { NavigateOptions, ToOptions } from './link'
 
 declare global {
   interface Window {
@@ -98,19 +100,8 @@ declare global {
   }
 }
 
-export interface Register {
-  // router: Router
-}
-
 export type AnyRouter = Router<any, any, any, any, any, any>
 
-export type AnyRouterWithContext<TContext> = Router<
-  AnyRouteWithContext<TContext>,
-  any,
-  any,
-  any,
-  any
->
 
 export type RegisteredRouter = Register extends {
   router: infer TRouter extends AnyRouter
