+++ packages/react-router/src/useLoaderData.tsx
@@ -4,9 +4,8 @@ import type {
   ValidateSelected,
 } from './structuralSharing'
 import type { AnyRouter, RegisteredRouter } from './router'
-import type { AllLoaderData, RouteById } from './routeInfo'
+import type { AllLoaderData, Expand, RouteById } from '@tanstack/router-core'
 import type { StrictOrFrom } from './utils'
-import type { Expand } from '@tanstack/router-core'
 
 export interface UseLoaderDataBaseOptions<
   TRouter extends AnyRouter,
