+++ packages/router-core/src/fileRoute.ts
@@ -1,0 +1,30 @@
+import type { AnyRoute, RootRoute } from './route'
+
+export interface FileRouteTypes {
+  fileRoutesByFullPath: any
+  fullPaths: any
+  to: any
+  fileRoutesByTo: any
+  id: any
+  fileRoutesById: any
+}
+
+export type InferFileRouteTypes<TRouteTree extends AnyRoute> =
+  TRouteTree extends RootRoute<
+    any,
+    any,
+    any,
+    any,
+    any,
+    any,
+    any,
+    infer TFileRouteTypes extends FileRouteTypes
+  >
+    ? TFileRouteTypes
+    : never
+
+export interface FileRoutesByPath {
+  // '/': {
+  //   parentRoute: typeof rootRoute
+  // }
+}
