+++ packages/solid-router/src/RouterProvider.tsx
@@ -1,60 +1,12 @@
 import { Matches } from './Matches'
 import { getRouterContext } from './routerContext'
 import type * as Solid from 'solid-js'
-import type { NavigateOptions, ToOptions } from './link'
-import type {
-  ParsedLocation,
-  ViewTransitionOptions,
-} from '@tanstack/router-core'
-import type { RoutePaths } from './routeInfo'
 import type {
   AnyRouter,
   RegisteredRouter,
   Router,
   RouterOptions,
 } from './router'
-
-export interface CommitLocationOptions {
-  replace?: boolean
-  resetScroll?: boolean
-  hashScrollIntoView?: boolean | ScrollIntoViewOptions
-  viewTransition?: boolean | ViewTransitionOptions
-  /**
-   * @deprecated All navigations use Solid transitions under the hood now
-   **/
-  startTransition?: boolean
-  ignoreBlocker?: boolean
-}
-
-export interface MatchLocation {
-  to?: string | number | null
-  fuzzy?: boolean
-  caseSensitive?: boolean
-  from?: string
-}
-
-export type NavigateFn = <
-  TRouter extends RegisteredRouter,
-  TTo extends string | undefined,
-  TFrom extends RoutePaths<TRouter['routeTree']> | string = string,
-  TMaskFrom extends RoutePaths<TRouter['routeTree']> | string = TFrom,
-  TMaskTo extends string = '',
->(
-  opts: NavigateOptions<TRouter, TFrom, TTo, TMaskFrom, TMaskTo>,
-) => Promise<void> | void
-
-export type BuildLocationFn = <
-  TRouter extends RegisteredRouter,
-  TTo extends string | undefined,
-  TFrom extends RoutePaths<TRouter['routeTree']> | string = string,
-  TMaskFrom extends RoutePaths<TRouter['routeTree']> | string = TFrom,
-  TMaskTo extends string = '',
->(
-  opts: ToOptions<TRouter, TFrom, TTo, TMaskFrom, TMaskTo> & {
-    leaveParams?: boolean
-    _includeValidateSearch?: boolean
-  },
-) => ParsedLocation
 
 export function RouterContextProvider<
   TRouter extends AnyRouter = RegisteredRouter,
