+++ packages/solid-router/src/fileRoute.ts
@@ -15,52 +15,24 @@ import type {
   AnyPathParams,
   AnyValidator,
   Constrain,
+  FileRoutesByPath,
   ResolveParams,
+  RouteById,
+  RouteIds,
 } from '@tanstack/router-core'
 
 import type {
   AnyRoute,
   FileBaseRouteOptions,
-  RootRoute,
   Route,
   RouteConstraints,
   RouteLoaderFn,
   UpdatableRouteOptions,
 } from './route'
 import type { RegisteredRouter } from './router'
-import type { RouteById, RouteIds } from './routeInfo'
 import type { UseLoaderDepsRoute } from './useLoaderDeps'
 import type { UseLoaderDataRoute } from './useLoaderData'
 import type { UseRouteContextRoute } from './useRouteContext'
-
-export interface FileRoutesByPath {
-  // '/': {
-  //   parentRoute: typeof rootRoute
-  // }
-}
-
-export interface FileRouteTypes {
-  fileRoutesByFullPath: any
-  fullPaths: any
-  to: any
-  fileRoutesByTo: any
-  id: any
-  fileRoutesById: any
-}
-
-export type InferFileRouteTypes<TRouteTree extends AnyRoute> =
-  TRouteTree extends RootRoute<
-    any,
-    any,
-    any,
-    any,
-    any,
-    any,
-    any,
-    infer TFileRouteTypes extends FileRouteTypes
-  >
-    ? TFileRouteTypes
-    : never
 
 export function createFileRoute<
   TFilePath extends keyof FileRoutesByPath,
