+++ packages/solid-router/src/not-found.tsx
@@ -2,7 +2,7 @@ import { CatchBoundary } from './CatchBoundary'
 import { useRouterState } from './useRouterState'
 import type * as Solid from 'solid-js'
 import type { RegisteredRouter } from './router'
-import type { RouteIds } from './routeInfo'
+import type { RouteIds } from '@tanstack/router-core'
 
 export type NotFoundError = {
   /**
