+++ packages/solid-router/src/router.ts
@@ -42,42 +42,43 @@ import type {
   AnyContext,
   AnySchema,
   AnyValidator,
+  BuildLocationFn,
   CommitLocationOptions,
   ControlledPromise,
+  FullSearchSchema,
+  MakeRemountDepsOptionsUnion,
   Manifest,
+  NavigateFn,
+  NavigateOptions,
   NonNullableUpdater,
   ParsedLocation,
   PickAsRequired,
+  Register,
   ResolveRelativePath,
+  RouteById,
+  RoutePaths,
+  RoutesById,
+  RoutesByPath,
   SearchMiddleware,
   SearchParser,
   SearchSerializer,
   StartSerializer,
+  ToOptions,
   TrailingSlashOption,
   Updater,
   ViewTransitionOptions,
 } from '@tanstack/router-core'
 import type {
   AnyRoute,
-  AnyRouteWithContext,
   BeforeLoadContextOptions,
   ErrorRouteComponent,
   LoaderFnContext,
-  MakeRemountDepsOptionsUnion,
   NotFoundRouteComponent,
   RootRoute,
   RouteComponent,
   RouteContextOptions,
   RouteMask,
 } from './route'
-
-import type {
-  FullSearchSchema,
-  RouteById,
-  RoutePaths,
-  RoutesById,
-  RoutesByPath,
-} from './routeInfo'
 import type {
   AnyRouteMatch,
   MakeRouteMatch,
@@ -85,11 +86,8 @@ import type {
   MatchRouteOptions,
 } from './Matches'
 
-import type { BuildLocationFn, NavigateFn } from './RouterProvider'
-
 import type { AnyRedirect, ResolvedRedirect } from './redirects'
 import type { NotFoundError } from './not-found'
-import type { NavigateOptions, ToOptions } from './link'
 
 declare global {
   interface Window {
@@ -97,19 +95,7 @@ declare global {
   }
 }
 
-export interface Register {
-  // router: Router
-}
-
 export type AnyRouter = Router<any, any, any, any, any>
-
-export type AnyRouterWithContext<TContext> = Router<
-  AnyRouteWithContext<TContext>,
-  any,
-  any,
-  any,
-  any
->
 
 export type RegisteredRouter = Register extends {
   router: infer TRouter extends AnyRouter
