+++ packages/solid-router/src/useSearch.tsx
@@ -1,10 +1,14 @@
 import { useMatch } from './useMatch'
 import type { ThrowConstraint } from './useMatch'
 import type { Accessor } from 'solid-js'
-import type { FullSearchSchema, RouteById } from './routeInfo'
 import type { AnyRouter, RegisteredRouter } from './router'
 import type { StrictOrFrom } from './utils'
-import type { Expand, ThrowOrOptional } from '@tanstack/router-core'
+import type {
+  Expand,
+  FullSearchSchema,
+  RouteById,
+  ThrowOrOptional,
+} from '@tanstack/router-core'
 
 export interface UseSearchBaseOptions<
   TRouter extends AnyRouter,
